recorrido completo desde controller hasta repositorio y devolviendo info al controller. (faltaria incorporar resultado a front)

+++ src/main/java/ar/edu/unlam/tallerweb1/domain/mascotas/ServicioMascota.java
@@ -3,5 +3,5 @@ package ar.edu.unlam.tallerweb1.domain.mascotas;
 import java.util.List;
 
 public interface ServicioMascota {
-    List<Mascota> ObtenerMascotasPorTipo(int idTipoMascota);
+    List<Mascota> ObtenerMascotasPorTipo(long idTipoMascota);
 }
